# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CRUD Barang",
        "description": "This is a sample crud barang with mongodb",
        "contact": {
            "name": "Ahmad Fauzi",
            "email": "fauzi442097@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/barang/{id}": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Find data barang",
                "description": "Find data barang in database by Id",
                "operationId": "0d5d3892027bd58b8c779a1b0d2c4dd5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": "200"
                                        },
                                        "message": {
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "namaBarang": {
                                                    "description": "Nama Barang",
                                                    "type": "string",
                                                    "maxLength": 100,
                                                    "example": "Meja"
                                                },
                                                "harga": {
                                                    "description": "Harga Barang",
                                                    "type": "number",
                                                    "example": "500000"
                                                },
                                                "stock": {
                                                    "description": "Stock Barang",
                                                    "type": "number",
                                                    "minLength": 1,
                                                    "example": "1"
                                                },
                                                "_id": {
                                                    "description": "Id Barang",
                                                    "type": "string",
                                                    "example": "62bfd45c1ac213e12405b513"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Response Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Barang"
                ],
                "summary": "Delete existing data barang",
                "description": "Delete existing data barang in database by Id",
                "operationId": "c4e24ce1b762ae40ee8a0043f0685076",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": "200"
                                        },
                                        "message": {
                                            "example": "Success deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Response Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "NotFoundResponse": {
                "title": "Response Not Found",
                "properties": {
                    "code": {
                        "type": "number",
                        "example": "404"
                    },
                    "message": {
                        "type": "string",
                        "example": "Data Not Found"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
...
            "id": {
                "name": "id",
                "in": "path",
                "description": "Id of data barang",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        }
    }
}